Complete front design with logic to get information about user
