+++ TripPlanner/TripPlanner.API/Dtos/Profile/ProfileInformationDto.cs
@@ -1,0 +1,7 @@
+﻿namespace TripPlanner.API.Dtos.Profile;
+
+public record ProfileInformationDto (
+    string Name,
+    string Surname,
+    string Email
+);
