+++ TripPlanner/TripPlanner.API/Services/Profile/IProfileService.cs
@@ -1,0 +1,9 @@
+﻿using TripPlanner.API.Dtos.Profile;
+
+namespace TripPlanner.API.Services.Profile
+{
+    public interface IProfileService
+    {
+        Task<ProfileInformationDto> GetUserInformation(string userId);
+    }
+}
